Update marker, add offline state notifier
