+++ map_system/static/map_system/local/application_options.js
@@ -34,9 +34,9 @@ var ApplicationOptions = {
     },
     leaflet: {
         iconUrls: {
-            normalIcon: '/static/map_system/images/markers/arrow_normal.png',
+            normalIcon: '/static/map_system/images/markers/car/red-top.png',
             alertedIcon: '/static/map_system/images/markers/arrow_alerted.png',
-            offlineIcon: '/static/map_system/images/markers/arrow_offline.png',
+            offlineIcon: '/static/map_system/images/markers/car/grey-top.png',
             warningIcon: '/static/map_system/images/markers/arrow_warning.png',
             defaultIcon: '/static/map_system/images/markers/car/red-top.png',
             resizeIcon: '/static/map_system/images/markers/resize.png'
