+++ map_system/templates/map_system/commons/header.html
@@ -90,6 +90,7 @@
     </script>
      <link rel="stylesheet" href="{% static 'map_system/libraries/app.css' %}">
 
+    <script src="{% static 'map_system/local/spatialObject.js' %}"></script>
     <script src="{% static 'map_system/local/websocket.js' %}"></script>
     <script src="{% static 'map_system/local/geo_remote.js' %}"></script>
     <script src="{% static 'map_system/local/geo_fencing.js' %}"></script>
